Remove parsing messages.
I've been experimenting for the past few days with different ways of
parsing Reddit's obnoxiously typed JSON into these protobuffers, and
concluded that the most coherenet and efficient way to do it is to
enumerate the field assignments manually. After discussing the
JSON->Proto issue with some others, I get the impression this is just
the standard way.

If I'm doing that (I am doing that), there is no point complicating the
messages with things that make them easier to generate. Most users will
be using graw to generate them and the messages will exist independent
of their source for much longer, so their utility in use should be
considered not just with more weight but in the stead of their ease in
writing from JSON.

A following commit will implement utilities for adapting Reddit JSON
responses to the protobuffers, and then I will consider this problem
solved. The messages must be stable before graw's api gaurantee.